move DropdownTab higher in the stack
