+++ ui/components/app/confirm-page-container/confirm-page-container-content/confirm-page-container-content.component.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import { Tabs, Tab, DropdownTab } from '../../../ui/tabs';
+import { Tabs, Tab } from '../../../ui/tabs';
 import Button from '../../../ui/button';
 import ActionableMessage from '../../../ui/actionable-message/actionable-message';
 import { PageContainerFooter } from '../../../ui/page-container';
@@ -75,29 +75,6 @@ export default class ConfirmPageContainerContent extends Component {
       insightComponent,
     } = this.props;
 
-    const insightSnaps = [
-      {
-        id: 'test snap',
-        manifest: {
-          proposedName: 'test snap',
-        },
-      },
-      {
-        id: 'Tenderly Insight',
-        manifest: {
-          proposedName: 'Tenderly Insight',
-        },
-      },
-    ];
-
-    const dropdownOptions = insightSnaps.reduce(
-      (prev, acc) => [
-        ...prev,
-        { value: acc.id, name: acc.manifest.proposedName },
-      ],
-      [],
-    );
-
     return (
       <Tabs>
         <Tab
@@ -117,14 +94,7 @@ export default class ConfirmPageContainerContent extends Component {
             {dataHexComponent}
           </Tab>
         )}
-        {insightComponent && (
-          <DropdownTab
-            className="confirm-page-container-content__tab"
-            options={dropdownOptions}
-          >
-            {insightComponent}
-          </DropdownTab>
-        )}
+        {insightComponent}
       </Tabs>
     );
   }
